requirements.txt
openai==1.52.2
pandas==2.0.3
tqdm==4.64.1
scikit-learn==1.3.1
python-dotenv==1.0.0
